Dynamically pulling the correct image from xkcd.
If a user provides a specific id (ie. thisapp.com/705) then we go to
xkcd and find out what picture is displayed for comic number 705.
We then pull that image and display it properly on the page.

Go ahead and try it, pick any id, it will show the same image as xkcd.

  http://localhost:4567/705
  http://xkcd.com/705