Using regular expressions to grab the current key.
There have been a number of changes to the xkcd homepage as of late.
The old strategy of key grabbing was not resistant to these changes.
The old strategy involved reading the page, parsing it with nokogiri,
then counting the number of h3s on the page. Once the desired h3 was
found, we would then read its contents.

This strategy is very frail when it comes to dynamic pages. So, I
thought about moving to a more resilient strategy. The two options
were to look for a nearby element and traverse the dom tree with
nokogiri, or to simply use regular expressions. I settled on using
regular expressions to solve this particular problem. This also
serves as a hat tip to the xkcd comic itself.
(See http://xkcd.com/208/)