Fix current comic id caching logic
We send a request to XKCD's homepage to determine the comic id that is
currently being displayed. Once we have that id, we do not need to go
back to their homepage over and over again because the comic id does
not update very frequently. Also, it's just polite. :) we don't want
to send any more traffic than we need to.

Any subsequent requests to our application that are still within the
freshness window will continue to use the cached value for the current
comic id, and will build a page based on the "relatively fresh" most
recent comic id.