Hide some implementation details in KeyGrabber
We prefer to use bare words in Ruby so we have the flexibility to
choose where the value of the bare word comes from. In this case, we
can reference the bare word `expire_at` and it can be a class-level
method, or a method level variable and it doesn't matter to the
caller.

Using bare words allows the expression `Time.now < expire_at` to
remain unchanged no matter where the value of the bare word
`expire_at` is coming from. I believe the majority of the benefit here
comes from the use of the bare words in general.

The two bare words `expire_at` and `last_read` were previously
accessible outside of the `KeyGrabber` class. This is not necessarily
problematic in general, however if they are accessible outside of the
class, then it introduces the possibility that another consumer could
begin to depend on the values and their implementation.

Since, there is not a compelling reason to expose the two values, and
as an experiment in code cleanliness, I am hiding the implementation
details of the bare words `expire_at` and `last_read` inside of the
`KeyGrabber` class. This gives us the flexibility to alter, rename, or
remove the bare words entirely with confidence that no one outside of
this class is using or depending on these values.