Navigation buttons remain in deterministic locations
As a user, it is confusing and unexpected when interface elements move
around the page, and especially so while they are currently in use.

Currently, while the user navigates from the current comic to any
previous comic the next button will show and hide. This is because on
the current comic page, there is no next comic.

When the next button appears and disappears it currently moves the
the previous button around the screen to make room for the presence or
absence of the next button.

Now, we preserve the positioning of the previous and next buttons no
matter if other navigation elements are present or absent. This makes
the user interface more deterministic, less confusing, and less error
prone especially when using the previous and next buttons in quick
succession to navigate between a number of comics, the buttons will
not move out from underneath the user while they are currently in use.