Rename TextGrabber to CaptionGrabber
The word `title` has been too overloaded in this project. It could
mean any one of, Browser Page Title, top level HTML h1 tag, or the
title attribute on the main image on the page.

This takes the latter example, the title attribute on the main image
on the page, and encapsulates it inside of the `CaptionGrabber` and
refers to it as `caption` elsewhere in the application.