Clean up RSpec warnings in image grabber spec
We should not be polluting the global namespace in our spec helper.
Instead of defining top level methods, we create an RSpecHelpers
module, and define our helper methods there.

Old RSpec syntax monkey-patched objects with `.should`, and this
behavior is no longer recommended. Now we use the new
`expect(foo).to eq()` syntax.

Instead of using FakeWeb, we are using WebMock