Handle new permalink structure
The XKCD comic page changed the way that it lists the permalink to the
current comic. It used to be a plaintext description and URL. Now the
URL is hyperlinked. This slightly alters the scraping syntax we need
to use.

The server is crashing right now, so we're for the time being, we just
copied the old regular expression logic and updated it to check for
the hyperlinked version as well. First, look for the old syntax, then
if no key is found, try the new syntax.